languages reducer test
