+++ app/reducers/index.js
@@ -1,7 +1,8 @@
 import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
 
-import { languages, targetLanguage } from './settings-reducer.js'
+import { languages } from './languages-reducer.js'
+import { targetLanguage } from './targetLanguage-reducer.js'
 import { currentCard } from './card-reducer.js'
 import { deck } from './carddeck-reducer.js'
 import { foundWords } from './imageUpload-reducer.js'
