+++ app/reducers/languages-reducer.js
@@ -1,0 +1,8 @@
+export const languages = (state = [], action) => {
+  switch (action.type) {
+    case 'ADD_LANGUAGES':
+      return [...action.languageSet]
+    default:
+      return state
+  }
+}
